% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffit.R
\name{diffit}
\alias{diffit}
\alias{print.diff}
\title{Aplication of the Diffit method.}
\usage{
diffit(datos, amb = 2, stand = TRUE, niter = 10000)
}
\arguments{
\item{datos}{datos original data from data frames}

\item{amb}{numbers of environment}

\item{stand}{a boolean value, if it is TRUE (value set by default) each
variable is centered and scale by variable.}

\item{niter}{the iteration number for the Tuckals algorithm, by default
10000 iteration.}
}
\value{
\code{saldiffit} a list with a combination numbers of axis,
percentage of variability explained and Diffit rate. The critic value or
threhold is also return.
}
\description{
The axis value to be retained in the first and second mode, P
and Q, was determined using the Diffit. For the third mode, K equal number
of environments. The Diffit method consist of fitting the Tuckals algorithm
in each combinations of axis value. The combination of axis are selected by
the Diffit criteria where \eqn{nc1<=nc2*nc3 && nc3<=nc1*nc2 && nc2 <= nc1*nc3}.
}
\details{
The best solution obtained is the one which combination obtained
has a higher porcentage of variability explained and it has to exceed the
threshold.
}
\examples{
data(maize_pop)
dif_sal <- diffit(maize_pop,amb=2)
print(dif_sal)

}
\author{
Marta Marticorena, Gustavo Gimenez, Cecilia Gonzalez, Sergio Bramardi
}
\references{
\describe{
 \item{MARTICORENA, M.; BRAMARDI, S.; DEFACIO, R. 2010.}{Characterization of
 maize populations in different environmental conditions by means of
 Three-Mode Principal Components Analysis. Revista Ciencia e Investigacion
 Agraria. 37(3): 93-105.}
 \item{Timmerman, M.E., and H. Kiers. 2000.}{Three-mode principal components
  analysis. Choosing numbers of components and sensitivity to local optima.
  The British Journal of the Mahematical and Statistical Psychology 53: 1-16.}
}
}

