% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrition.R
\name{matrition}
\alias{matrition}
\title{Concatenate data frame in array and matrix by cases, variables and
environments}
\usage{
matrition(datos, I, J, K)
}
\arguments{
\item{datos}{original data from data frames}

\item{I}{the numbers of cases}

\item{J}{the numbers of variables}

\item{K}{the numbers of environment or conditions}
}
\value{
\code{matrizlista} return a list with: the array "m" with all the
data concatenate in array. X1 the data is concatenate by cases, X2  the data
concatenate by variables and X3 the data concatenate by environments.
}
\description{
Concatenate data frame in array and matrix by cases, variables
 and environments to performs three mode principal components with the
 function \code{\link{tucker2R}}.
}
\details{
This process is also knowing as 'matricizing'  or 'unfolding'.
}
\examples{
data(maize_pop,package = "tuckerR.mmgg")
conc_matrix <- matrition(maize_pop,I=30,J=10,K=2)
conc_matrix$m  #get m array
conc_matrix$X1 #get matrix by cases
conc_matrix$X2 #get matrix by variables
conc_matrix$X3 #get matrix by environments
}
