% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuckerR.mmgg.R
\docType{package}
\name{tuckerR.mmgg}
\alias{tuckerR.mmgg}
\alias{tuckerR.mmgg-package}
\title{Three-Mode Principal Components: Tuckers  Models}
\description{
This package performs Three-Mode Principal Components  using Tuckers  Models
and  plot  interactive  Biplot. These Three-Mode Principal Component Analysis,
provides both useful analytic and graphic tools to study and characterize
phytogenetic resources, specially when the influence of environmental
factors are possible.
}
\author{
Marta Marticorena, Gustavo Gimenez, Cecilia Gonzalez, Sergio Bramardi
}

