% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffit.R
\name{diffit}
\alias{diffit}
\alias{print.diff}
\title{Aplication of the Diffit method.}
\usage{
diffit(datos, amb = 2, stand = TRUE, niter = 10000)
}
\arguments{
\item{datos}{datos original data from data frames}

\item{amb}{numbers of environment}

\item{stand}{a boolean value, if it is TRUE (value set by default) each
variable is centered and scale by variable.}

\item{niter}{the iteration number for the Tuckals algorithm, by default
10000 iteration.}
}
\value{
\code{saldiffit} a list with a combination numbers of axis,
percentage of variability explained and Diffit rate. The critic value or
threhold is also return.
}
\description{
The diffit method is used to apply when we need to know the
axis number to be gathered in the P mode, and Q mode. The third mode, K
it is related to the environment  numbers.
The diffit method consist on fitting each value with the Tuckle algorithm.
}
\details{
The final result is the model which has the most coefficient diffits
the greatest variability explained and the one which exceed the threshold.
}
\examples{
#Copy and paste this example in your console without the comment
#data(maize_pop,package = "tuckerR.mmgg")
#dif_sal <- diffit(maize_pop,amb=2)
#print(dif_sal) the best combination is 3 3 2

}
\author{
Marta Marticorena, Gustavo Gimenez, Cecilia Gonzalez, Sergio Bramardi
}
\references{
\describe{
 \item{MARTICORENA, M.; BRAMARDI, S.; DEFACIO, R. 2010.}{Characterization of
 maize populations in different environmental conditions by means of
 Three-Mode Principal Components Analysis. Revista Ciencia e Investigacion
 Agraria. 37(3): 93-105.}
 \item{Timmerman, M.E., and H. Kiers. 2000.}{Three-mode principal components
  analysis. Choosing numbers of components and sensitivity to local optima.
  The British Journal of the Mahematical and Statistical Psychology 53: 1-16.}
}
}

